requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
autopep8==1.6.0
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Django==4.1.4
django-cors-headers==3.14.0
django-otp==1.1.4
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.4
mysql-connector-python==8.0.32
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.1.1
prompt-toolkit==3.0.38
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.0.0
pypng==0.20220715.0
pytz==2022.7.1
qrcode==7.4.2
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.16.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
zope.interface==5.4.0
